Create EditorHelper.js with a reusable copy function
